+++ application/src/app/(app)/saved.tsx
@@ -1,5 +1,4 @@
 import { Ionicons, MaterialCommunityIcons } from '@expo/vector-icons';
-import { Image } from 'expo-image';
 import * as Linking from 'expo-linking';
 import React, { useCallback } from 'react';
 import { Pressable } from 'react-native';
@@ -72,50 +71,52 @@ function SocialPostCard({ post }: { post: SocialPost }) {
   }, [post.source_link]);
 
   return (
-    <View className="relative mb-4 overflow-hidden rounded-2xl border border-neutral-100 bg-white p-5 shadow-sm dark:border-neutral-800 dark:bg-neutral-900">
-      {/* Decorative Quote Icon */}
-      <View className="absolute -right-4 -top-4 opacity-5">
-        <MaterialCommunityIcons
-          name="format-quote-close"
-          size={80}
-          color="black"
-        />
-      </View>
-
-      <View className="mb-3 flex-row items-center gap-3">
-        <View className="size-10 items-center justify-center rounded-full bg-neutral-100 dark:bg-neutral-800">
-          <Text className="text-lg font-bold text-neutral-700 dark:text-neutral-300">
-            {post.name.charAt(0)}
-          </Text>
+    <View className="mb-4 rounded-xl border border-neutral-200 bg-white shadow-md">
+      <View className="relative overflow-hidden p-5">
+        {/* Decorative Quote Icon */}
+        <View className="absolute -right-4 -top-4 opacity-5">
+          <MaterialCommunityIcons
+            name="format-quote-close"
+            size={80}
+            color="black"
+          />
         </View>
-        <View className="flex-1">
-          <Text className="text-base font-bold text-neutral-900 dark:text-white">
-            {post.name}
-          </Text>
-          <Text className="text-xs font-medium text-neutral-400">
-            {formatRelativeTime(post.created_at)}
-          </Text>
+
+        <View className="mb-3 flex-row items-center gap-3">
+          <View className="size-10 items-center justify-center rounded-full bg-neutral-100 dark:bg-neutral-800">
+            <Text className="text-lg font-bold text-neutral-700 dark:text-neutral-300">
+              {post.name.charAt(0)}
+            </Text>
+          </View>
+          <View className="flex-1">
+            <Text className="text-base font-bold text-neutral-900 dark:text-white">
+              {post.name}
+            </Text>
+            <Text className="text-xs font-medium text-neutral-400">
+              {formatRelativeTime(post.created_at)}
+            </Text>
+          </View>
         </View>
+
+        <Text className="mb-4 text-base leading-7 text-neutral-600 dark:text-neutral-300">
+          {post.description}
+        </Text>
+
+        {post.source_link && (
+          <Pressable
+            onPress={handleOpenSource}
+            className="flex-row items-center self-start gap-2 rounded-full border border-blue-100 bg-blue-50 px-4 py-2 active:opacity-70 dark:border-blue-900/30 dark:bg-blue-900/20"
+          >
+            <Ionicons name="link" size={14} color="#3b82f6" />
+            <Text className="text-xs font-medium text-neutral-500 dark:text-neutral-400">
+              Source
+            </Text>
+            <Text className="text-xs font-bold text-blue-600 dark:text-blue-400">
+              {post.source_name}
+            </Text>
+          </Pressable>
+        )}
       </View>
-
-      <Text className="mb-4 text-base leading-7 text-neutral-600 dark:text-neutral-300">
-        {post.description}
-      </Text>
-
-      {post.source_link && (
-        <Pressable
-          onPress={handleOpenSource}
-          className="self-start flex-row items-center gap-2 rounded-full border border-blue-100 bg-blue-50 px-4 py-2 dark:border-blue-900/30 dark:bg-blue-900/20 active:opacity-70"
-        >
-          <Ionicons name="link" size={14} color="#3b82f6" />
-          <Text className="text-xs font-medium text-neutral-500 dark:text-neutral-400">
-            Source
-          </Text>
-          <Text className="text-xs font-bold text-blue-600 dark:text-blue-400">
-            {post.source_name}
-          </Text>
-        </Pressable>
-      )}
     </View>
   );
 }
@@ -127,8 +128,8 @@ export default function SocialPosts() {
       edges={['top']}
     >
       <FocusAwareStatusBar />
-      <View className="flex-1 px-5 pt-6">
-        <View className="mb-8 flex-row items-center justify-between">
+      <View className="flex-1 pt-6">
+        <View className="flex-row items-center justify-between px-5 pb-4 border-b border-neutral-200 shadow-sm bg-white dark:bg-neutral-950 dark:border-neutral-800">
           <View className="mr-4 flex-1">
             <Text className="text-3xl font-black text-neutral-900 dark:text-white">
               Social posts
@@ -138,16 +139,17 @@ export default function SocialPosts() {
               curated from professional networks like LinkedIn and Twitter.
             </Text>
           </View>
-          <View className="rounded-2xl p-2">
-            <Image
-              source="https://cdn-icons-png.flaticon.com/512/4207/4207285.png"
-              style={{ width: 44, height: 44 }}
-              contentFit="contain"
-            />
-          </View>
         </View>
 
-        <ScrollView className="flex-1" showsVerticalScrollIndicator={false}>
+        <ScrollView
+          className="flex-1"
+          showsVerticalScrollIndicator={false}
+          contentContainerStyle={{
+            paddingHorizontal: 20,
+            paddingBottom: 20,
+            paddingTop: 10,
+          }}
+        >
           {DUMMY_POSTS.map((post) => (
             <SocialPostCard key={post.id} post={post} />
           ))}
